# schemas/resonance_api_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mothership Cockpit",
    "description": "\n# Mothership Cockpit API\n\nLocal integration backend for the Mothership Cockpit system.\n\n## Features\n\n- **Session Management**: Create and manage user sessions\n- **Operation Tracking**: Track long-running operations with progress\n- **Component Health**: Monitor system component health\n- **Alert System**: Create, acknowledge, and resolve alerts\n- **Real-time Updates**: WebSocket support for live updates\n- **Cloud Integration**: Seamless Gemini Studio integration\n\n## Authentication\n\nThe API supports multiple authentication methods:\n- API Key (X-API-Key header)\n- JWT Bearer Token (Authorization header)\n\nDevelopment mode allows unauthenticated access.\n        ",
    "version": "1.0.0",
    "x-resonance-api-version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/api-keys/": {
      "get": {
        "tags": [
          "api-keys"
        ],
        "summary": "List Api Keys",
        "description": "List all API keys for the current user.",
        "operationId": "list_api_keys_api_v1_api_keys__get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Auth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyResponse"
                  },
                  "type": "array",
                  "title": "Response List Api Keys Api V1 Api Keys  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "api-keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key.\n\nThe plaintext key is only shown once - make sure to store it securely.",
        "operationId": "create_api_key_api_v1_api_keys__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_api_key_api_v1_api_keys__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-keys/{key_id}": {
      "get": {
        "tags": [
          "api-keys"
        ],
        "summary": "Get Api Key",
        "description": "Get API key details.",
        "operationId": "get_api_key_api_v1_api_keys__key_id__get",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Auth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "api-keys"
        ],
        "summary": "Update Api Key",
        "description": "Update an API key (name or active status).",
        "operationId": "update_api_key_api_v1_api_keys__key_id__patch",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_api_key_api_v1_api_keys__key_id__patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "api-keys"
        ],
        "summary": "Delete Api Key",
        "description": "Delete an API key.",
        "operationId": "delete_api_key_api_v1_api_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Auth"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreatedResponse": {
        "properties": {
          "api_key": {
            "$ref": "#/components/schemas/APIKeyResponse"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The actual API key (only shown once, store securely)"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "key"
        ],
        "title": "APIKeyCreatedResponse",
        "description": "Response after creating an API key (includes the key itself - only shown once)."
      },
      "APIKeyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "key_prefix",
          "name",
          "tier",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "APIKeyResponse",
        "description": "API key response (without sensitive hash)."
      },
...
      "Body_create_api_key_api_v1_api_keys__post": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/CreateAPIKeyRequest"
          },
          "auth": {
            "additionalProperties": true,
            "type": "object",
            "title": "Auth"
          }
        },
        "type": "object",
        "required": [
          "request"
        ],
        "title": "Body_create_api_key_api_v1_api_keys__post"
      },
...
      "Body_update_api_key_api_v1_api_keys__key_id__patch": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/UpdateAPIKeyRequest"
          },
          "auth": {
            "additionalProperties": true,
            "type": "object",
            "title": "Auth"
          }
        },
        "type": "object",
        "required": [
          "request"
        ],
        "title": "Body_update_api_key_api_v1_api_keys__key_id__patch"
      },
...
      "CreateAPIKeyRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "User-friendly name for the API key"
          },
          "expires_in_days": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Days",
            "description": "Number of days until expiration (None for no expiration)"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateAPIKeyRequest",
        "description": "Request to create an API key."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateAPIKeyRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Update the name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "Activate or deactivate the key"
          }
        },
        "type": "object",
        "title": "UpdateAPIKeyRequest",
        "description": "Request to update an API key."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "API key for authentication",
        "in": "header",
        "name": "X-API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "description": "JWT Bearer token",
        "scheme": "bearer"
      }
    }
  }
}